pyproject.toml
[project]
name = "open-science"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "fake-useragent>=2.2.0",
    "fastapi>=0.123.9",
    "grobid-client-python>=0.1.2",
    "lxml>=6.0.2",
    "numpy>=2.3.5",
    "playwright>=1.56.0",
    "pypdf>=6.4.0",
    "selenium>=4.38.0",
    "selenium-driverless>=1.9.4",
    "uvicorn>=0.38.0",
    "webdriver-manager>=4.0.2",
]
